refactor common postprocess functions
